refactor: exception, delete API 수정
